requirements.txt
Django==5.2.8
asgiref==3.10.0
sqlparse==0.5.3
openai==1.51.2
Pillow==10.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
transformers==4.46.3
torch==2.5.1
scikit-learn==1.6.0
opencv-python
tensorflow
fer
mediapipe